+++ shadow/app/dashboard/page.tsx
@@ -14,39 +14,40 @@ export default function DashboardPage() {
   const { activeId } = useDashboardActive();
 
   return (
-    <div className='flex h-full w-full'>
-      {/* Left Panel - Dashboard Report */}
-      <div className='flex-1 flex flex-col'>
-        {/* Top Section */}
-        <DashboardHeader />
+    // <div className='flex h-full w-full'>
+    //   {/* Left Panel - Dashboard Report */}
+    //   <div className='flex-1 flex flex-col'>
+    //     {/* Top Section */}
+    //     <DashboardHeader />
 
-        {/* Bottom Section - Scrollable */}
-        <div className='flex-1 overflow-y-auto p-4'>
-          <div className='bg-gray-50 rounded-lg min-h-full'>
-            <DashboardCanvas
-              dashboardSections={sections}
-              removeSection={remove}
-            />
-          </div>
-        </div>
-      </div>
+    //     {/* Bottom Section - Scrollable */}
+    //     <div className='flex-1 overflow-y-auto p-4'>
+    //       <div className='bg-gray-50 rounded-lg min-h-full'>
+    //         <DashboardCanvas
+    //           dashboardSections={sections}
+    //           removeSection={remove}
+    //         />
+    //       </div>
+    //     </div>
+    //   </div>
 
-      {/* Right Panel - Block Settings */}
-      {activeId && (
-        <div className='w-96 border-l'>
-          <VisualizationSetting onUpdateSection={update} />
-        </div>
-      )}
+    //   {/* Right Panel - Block Settings */}
+    //   {activeId && (
+    //     <div className='w-96 border-l'>
+    //       <VisualizationSetting onUpdateSection={update} />
+    //     </div>
+    //   )}
 
-      {/* Toolbar */}
-      <div className='fixed bottom-4 left-1/2 transform -translate-x-1/2 bg-white shadow-lg rounded-lg px-4 py-2 border'>
-        <DashboardToolbar
-          onAddBlock={add}
-          onAddVariable={() => {}}
-          onAddHeader={() => {}}
-          onAddText={() => {}}
-        />
-      </div>
-    </div>
+    //   {/* Toolbar */}
+    //   <div className='fixed bottom-4 left-1/2 transform -translate-x-1/2 bg-white shadow-lg rounded-lg px-4 py-2 border'>
+    //     <DashboardToolbar
+    //       onAddBlock={add}
+    //       onAddVariable={() => {}}
+    //       onAddHeader={() => {}}
+    //       onAddText={() => {}}
+    //     />
+    //   </div>
+    // </div>
+    <></>
   );
 }